Delete Destroyer and Upserter; Reorganize structure
