+++ src/Integration/Laravel/Receive/Savers/ConflictResolvers/ByTargetKeysResolver.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Umbrellio\TableSync\Integration\Laravel\Receive\Upserter;
+namespace Umbrellio\TableSync\Integration\Laravel\Receive\Savers\ConflictResolvers;
 
 use Umbrellio\TableSync\Integration\Laravel\Receive\MessageData\MessageData;
