+++ src/Integration/Laravel/TableSyncServiceProvider.php
@@ -12,8 +12,8 @@ use Psr\Log\LoggerInterface;
 use Umbrellio\TableSync\Integration\Laravel\Console\PidManager;
 use Umbrellio\TableSync\Integration\Laravel\Publishers\QueuePublisher;
 use Umbrellio\TableSync\Integration\Laravel\Receive\MessageData\MessageDataRetriever;
-use Umbrellio\TableSync\Integration\Laravel\Receive\Upserter\ByTargetKeysResolver;
-use Umbrellio\TableSync\Integration\Laravel\Receive\Upserter\ConflictConditionResolverContract;
+use Umbrellio\TableSync\Integration\Laravel\Receive\Savers\ConflictResolvers\ByTargetKeysResolver;
+use Umbrellio\TableSync\Integration\Laravel\Receive\Savers\ConflictResolvers\ConflictConditionResolverContract;
 use Umbrellio\TableSync\Integration\Laravel\ReceivedMessageHandlers\QueueReceivedMessageHandler;
 use Umbrellio\TableSync\Publisher;
 use Umbrellio\TableSync\Rabbit\ChannelContainer;
