+++ tests/functional/Laravel/Integration/Receive/Savers/ConflictResolvers/ByTargetKeysResolverTest.php
@@ -2,12 +2,12 @@
 
 declare(strict_types=1);
 
-namespace Umbrellio\TableSync\Tests\functional\Laravel\Integration\Receive\Upserter;
+namespace Umbrellio\TableSync\Tests\functional\Laravel\Integration\Receive\Savers\ConflictResolvers;
 
 use Illuminate\Support\Facades\App;
 use Umbrellio\TableSync\Integration\Laravel\Receive\MessageData\MessageData;
+use Umbrellio\TableSync\Integration\Laravel\Receive\Savers\ConflictResolvers\ByTargetKeysResolver;
 use Umbrellio\TableSync\Integration\Laravel\Receive\Savers\QuerySaver;
-use Umbrellio\TableSync\Integration\Laravel\Receive\Upserter\ByTargetKeysResolver;
 use Umbrellio\TableSync\Tests\functional\Laravel\LaravelTestCase;
 
 class ByTargetKeysResolverTest extends LaravelTestCase
